+++ src/assets/styles/index.sass
@@ -8,7 +8,10 @@
 
 // Variables
 @import './partials/_variables'
+@import './partials/_breakpoints'
+@import './partials/_mixins'
+@import './partials/_global'
 
 body
-    font-family: $PatuaOne
+    font-family: $Poppins
     overflow: hidden
