+++ src/assets/styles/partials/_breakpoints.sass
@@ -1,19 +1,50 @@
-@mixin for-desktop
-    @media (min-width: 1200px)
+// ** for Min width
+@mixin for-min-xxl
+    @media (min-width: $breakpoint-xxl)
         @content
 
-@mixin for-laptop
-    @media (min-width: 1024px)
+@mixin for-min-xl
+    @media (min-width: $breakpoint-xl)
         @content
 
-@mixin for-tablet-landscape
-    @media (min-width: 900px)
+@mixin for-min-lg
+    @media (min-width: $breakpoint-lg)
         @content
 
-@mixin for-tablet
-    @media (min-width: 768px)
+@mixin for-min-md
+    @media (min-width: $breakpoint-md)
         @content
 
-@mixin for-mobile-large
-    @media (min-width: 425px)
+@mixin for-min-sm
+    @media (min-width: $breakpoint-sm)
         @content
+
+@mixin for-min-xs
+    @media (min-width: $breakpoint-xs)
+        @content
+
+// ** for Max width
+@mixin for-max-xxl
+    @media (max-width: $breakpoint-xxl)
+        @content
+
+@mixin for-max-xl
+    @media (max-width: $breakpoint-xl)
+        @content
+
+@mixin for-max-lg
+    @media (max-width: $breakpoint-lg)
+        @content
+
+@mixin for-max-md
+    @media (max-width: $breakpoint-md)
+        @content
+
+@mixin for-max-sm
+    @media (max-width: $breakpoint-sm)
+        @content
+
+@mixin for-max-xs
+    @media (max-width: $breakpoint-xs)
+        @content
+
