+++ src/assets/styles/partials/_global.sass
@@ -1,0 +1,14 @@
+.d-none
+    display: none
+.d-flex
+    display: flex
+.d-block
+    display: block
+.justify-between
+    justify-content: space-between
+.justify-around
+    justify-content: space-around
+.justify-center
+    justify-content: center
+.align-items-center
+    align-items: center
