+++ src/assets/styles/partials/_mixins.sass
@@ -1,0 +1,4 @@
+@mixin flex-center
+    display: flex
+    align-items: center
+    justify-content: center
