+++ src/assets/styles/partials/_variables.sass
@@ -1,9 +1,39 @@
-// Colors
-$orange: #FF9F46
-$white: #ffffff
-$blackStrong: #151515
-$blackLight: #1a1a1a
+@import url('https://fonts.googleapis.com/css2?family=Patua+One&display=swap')
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap')
 
 // Font family
-@import url('https://fonts.googleapis.com/css2?family=Patua+One&display=swap')
 $PatuaOne: 'Patua One', cursive
+$Poppins: 'Poppins', sans-serif
+
+// ** Colors
+$blue: #00cfe8
+$indigo: #6610f2
+$purple: #7367f0
+$pink: #e83e8c
+$red: #ea5455
+$orange: #ff9f43
+$orange-strong: #F58634
+$yellow: #ffc107
+$green: #28c76f
+$teal: #20c997
+$cyan: #17a2b8
+$white: #fff
+$gray: #b8c2cc
+$gray-dark: #1e1e1e
+$primary: #377dc3
+$secondary: #82868b
+$success: #28c76f
+$info: #00cfe8
+$warning: #ff9f43
+$danger: #ea5455
+$light: #f6f6f6
+$light-strong: #f3f3f3
+$dark: #4b4b4b
+
+// ** Breakpoints
+$breakpoint-xs: 0
+$breakpoint-sm: 576px
+$breakpoint-md: 768px
+$breakpoint-lg: 992px
+$breakpoint-xl: 1200px
+$breakpoint-xxl: 1440px
