+++ src/components/Inicio/_inicio.sass
@@ -7,104 +7,104 @@
     margin-top: -3rem
     justify-content: center
 
-    @include for-tablet-landscape
-        flex-direction: row
-        justify-content: space-around
+    // @include for-tablet-landscape
+    //     flex-direction: row
+    //     justify-content: space-around
 
-    .Marketing
-        display: flex
-        flex-direction: column
-        align-items: center
-        padding: 10px 1.5rem
-        gap: 10px
+    // .Marketing
+    //     display: flex
+    //     flex-direction: column
+    //     align-items: center
+    //     padding: 10px 1.5rem
+    //     gap: 10px
 
-        @include for-tablet-landscape
-            width: 50%
-            margin-left: 1rem
-            align-items: flex-start
-            text-align: start
+    //     @include for-tablet-landscape
+    //         width: 50%
+    //         margin-left: 1rem
+    //         align-items: flex-start
+    //         text-align: start
 
-        .Characteristics
-            strong
-                font-size: 15px
-            .faCircle
-                color: white
-                font-size: 6px
+    //     .Characteristics
+    //         strong
+    //             font-size: 15px
+    //         .faCircle
+    //             color: white
+    //             font-size: 6px
 
-            @include for-mobile-large
-                strong
-                    font-size: 17px
-                .faCircle
-                    font-size: 7px
+    //         @include for-mobile-large
+    //             strong
+    //                 font-size: 17px
+    //             .faCircle
+    //                 font-size: 7px
 
-            @include for-tablet
-                strong
-                    font-size: 18px
-                    transition: .2s all
+    //         @include for-tablet
+    //             strong
+    //                 font-size: 18px
+    //                 transition: .2s all
 
-            @include for-tablet-landscape
-                strong
-                    font-size: 20px
-                .faCircle
-                    font-size: 9px
+    //         @include for-tablet-landscape
+    //             strong
+    //                 font-size: 20px
+    //             .faCircle
+    //                 font-size: 9px
 
-            @include for-desktop
-                gap: 1rem
-                strong
-                    font-size: 24px
+    //         @include for-desktop
+    //             gap: 1rem
+    //             strong
+    //                 font-size: 24px
 
-        .Description
-            span
-                font-size: 14px
+    //     .Description
+    //         span
+    //             font-size: 14px
 
-                @include for-mobile-large
-                    font-size: 16px
+    //             @include for-mobile-large
+    //                 font-size: 16px
 
-                @include for-tablet
-                    font-size: 18px
+    //             @include for-tablet
+    //                 font-size: 18px
 
-                @include for-tablet-landscape
-                    font-size: 20px
+    //             @include for-tablet-landscape
+    //                 font-size: 20px
 
-                @include for-desktop
-                    font-size: 24px
+    //             @include for-desktop
+    //                 font-size: 24px
 
-        button
-            font-family: $PatuaOne
-            font-size: 12px
-            padding: 7px 14px
+    //     button
+    //         font-family: $PatuaOne
+    //         font-size: 12px
+    //         padding: 7px 14px
 
-            @include for-mobile-large
-                font-size: 14px
+    //         @include for-mobile-large
+    //             font-size: 14px
 
-            @include for-tablet
-                font-size: 15px
+    //         @include for-tablet
+    //             font-size: 15px
 
-            @include for-desktop
-                font-size: 20px
-                padding: 8px 20px
-                border-radius: 8px
+    //         @include for-desktop
+    //             font-size: 20px
+    //             padding: 8px 20px
+    //             border-radius: 8px
 
-        @include for-tablet
-            gap: 1rem
+    //     @include for-tablet
+    //         gap: 1rem
 
-    .BurguerImg
-        img
-            max-width: 500px
-            width: 250px
-            transition: .2s
+    // .BurguerImg
+    //     img
+    //         max-width: 500px
+    //         width: 250px
+    //         transition: .2s
 
-            @include for-mobile-large
-                width: 300px
+    //         @include for-mobile-large
+    //             width: 300px
 
-            @include for-tablet
-                width: 320px
+    //         @include for-tablet
+    //             width: 320px
 
-            @include for-laptop
-                width: 350px
+    //         @include for-laptop
+    //             width: 350px
 
-            @include for-desktop
-                width: 420px
+    //         @include for-desktop
+    //             width: 420px
 
-        @include for-tablet-landscape
-            width: 40%
+    //     @include for-tablet-landscape
+    //         width: 40%
